+++ tests/conftest.py
@@ -1,43 +1,83 @@
 import pytest
+from slugify import slugify
 
-from typing import Generator
 from werkzeug.security import generate_password_hash
 
-from tortoise.contrib.fastapi import register_tortoise
-from tortoise.contrib.test import initializer, finalizer
+from sqlalchemy_utils import create_database
+from alembic.command import upgrade
+from alembic.config import Config
 from fastapi.testclient import TestClient
 
+from api.polls.models import Poll
 from api.users.models import User
-from core.settings import TORTOISE_MODELS_LIST
+from api.auth.authorization import get_access_token
 from main import app
+from core.base import Base
+from core.database import engine, SessionLocal
+from core.settings import DATABASE_URI
 
+try:
+    create_database(engine.url)
+except Exception:
+    print('database exists')
 
-@pytest.yield_fixture(autouse=True)
-def initialize_tests(request):
-    register_tortoise(
-        app,
-        db_url="sqlite://:memory:",
-        modules={"models": TORTOISE_MODELS_LIST},
-        generate_schemas=True,
-        add_exception_handlers=True,
-    )
-    initializer(TORTOISE_MODELS_LIST)
-    request.addfinalizer(finalizer)
-
-
-@pytest.fixture(scope="session")
-def client() -> Generator:
-    yield TestClient(app)
+alembic_config = Config('alembic.ini')
+alembic_config.set_main_option('sqlalchemy.url', DATABASE_URI)
+upgrade(alembic_config, 'head')
 
 
 @pytest.fixture()
-async def user():
-    user = await User.create(username='test_user', email='test@mail.com',
-                             password=generate_password_hash('testpass123', method='sha256'))
+def db():
+    try:
+        db = SessionLocal()
+        yield db
+    finally:
+        db.close()
+
+
+@pytest.yield_fixture(autouse=True)
+def tables():
+    Base.metadata.create_all(engine)
+    yield
+    Base.metadata.drop_all(engine)
+
+
+@pytest.fixture(scope="session")
+def client() -> TestClient:
+    yield TestClient(app)
+
+
+@pytest.fixture
+def poll(active_user, db):
+    title = "new test poll"
+    poll = Poll(title=title, description="Something about test poll", slug=slugify(title),
+                creator=active_user)
+    db.add(poll)
+    db.commit()
+    db.refresh(poll)
+    return poll
+
+
+@pytest.fixture()
+def user(db):
+    user = User(username='non_active_user', email='non_active_user@mail.com',
+                password=generate_password_hash('testpass123', method='sha256'))
+    db.add(user)
+    db.commit()
+    db.refresh(user)
     yield user
 
 
 @pytest.fixture()
-async def active_user():
-    yield await User.create(username='test_user', is_active=True, email='test@mail.com',
-                            password=generate_password_hash('testpass123', method='sha256'))
+def token_header(active_user):
+    return {'Authorization': f"JWT {get_access_token(active_user)}"}
+
+
+@pytest.fixture()
+def active_user(db):
+    user = User(username='active_user', is_active=True, email='active_user@mail.com',
+                password=generate_password_hash('testpass123', method='sha256'))
+    db.add(user)
+    db.commit()
+    db.refresh(user)
+    return user